Remove instance references filter
